requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==3.1.1
    # via google-auth
certifi==2018.11.29
    # via requests
charset-normalizer==2.0.6
    # via requests
google-api-core[grpc]==1.31.2
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-monitoring
    #   google-cloud-trace
    #   opencensus
google-api-python-client==1.10.0
    # via google-cloud-profiler
google-auth==1.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-profiler
google-auth-httplib2==0.0.4
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-cloud-core==1.7.2
    # via google-cloud-trace
google-cloud-monitoring==2.5.0
    # via opencensus-ext-stackdriver
google-cloud-profiler==3.0.5
    # via -r requirements.in
google-cloud-trace==0.24.0
    # via opencensus-ext-stackdriver
googleapis-common-protos==1.52.0
    # via google-api-core
grpcio==1.39.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpcio-health-checking
grpcio-health-checking==1.39.0
    # via -r requirements.in
httplib2==0.18.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.10
    # via requests
opencensus==0.7.13
    # via
    #   -r requirements.in
    #   opencensus-ext-stackdriver
opencensus-context==0.1.2
    # via opencensus
opencensus-ext-stackdriver==0.8.0
    # via -r requirements.in
packaging==21.0
    # via
    #   google-api-core
    #   google-cloud-monitoring
proto-plus==1.19.0
    # via google-cloud-monitoring
protobuf==3.12.2
    # via
    #   google-api-core
    #   google-cloud-profiler
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   proto-plus
pyasn1==0.4.5
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.4
    # via google-auth
pyparsing==2.4.7
    # via packaging
python-json-logger==0.1.9
    # via -r requirements.in
pytz==2018.9
    # via google-api-core
requests==2.26.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   grpcio
    #   protobuf
uritemplate==3.0.0
    # via google-api-python-client
urllib3==1.24.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
